Move bnsSwapQueryTags into tags.ts
