+++ packages/iov-bns/src/client.spec.ts
@@ -28,7 +28,7 @@ import { TxQuery } from "@iov/tendermint-types";
 
 import { bnsCodec } from "./bnscodec";
 import { Client } from "./client";
-import { bnsFromOrToTag } from "./tags";
+import { bnsFromOrToTag, bnsSwapQueryTags } from "./tags";
 import { keyToAddress } from "./util";
 
 const skipTests = (): boolean => !process.env.BOV_ENABLED;
@@ -530,19 +530,19 @@ describe("Integration tests with bov+tendermint", () => {
     // ----- client.searchTx() -----
     // we should be able to find the transaction through quite a number of tag queries
 
-    const txById = await client.searchTx({ tags: [Client.swapQueryTags(querySwapId)] });
+    const txById = await client.searchTx({ tags: [bnsSwapQueryTags(querySwapId)] });
     expect(txById.length).toEqual(1);
     expect(txById[0].txid).toEqual(txid);
 
-    const txBySender = await client.searchTx({ tags: [Client.swapQueryTags(querySwapSender)] });
+    const txBySender = await client.searchTx({ tags: [bnsSwapQueryTags(querySwapSender)] });
     expect(txBySender.length).toEqual(1);
     expect(txBySender[0].txid).toEqual(txid);
 
-    const txByRecipient = await client.searchTx({ tags: [Client.swapQueryTags(querySwapRecipient)] });
+    const txByRecipient = await client.searchTx({ tags: [bnsSwapQueryTags(querySwapRecipient)] });
     expect(txByRecipient.length).toEqual(1);
     expect(txByRecipient[0].txid).toEqual(txid);
 
-    const txByHash = await client.searchTx({ tags: [Client.swapQueryTags(querySwapHash)] });
+    const txByHash = await client.searchTx({ tags: [bnsSwapQueryTags(querySwapHash)] });
     expect(txByHash.length).toEqual(1);
     expect(txByHash[0].txid).toEqual(txid);
