+++ packages/iov-bns/src/client.ts
@@ -41,9 +41,9 @@ import { ChainId, PostableBytes, Tag, TxId, TxQuery } from "@iov/tendermint-type
 import { bnsCodec } from "./bnscodec";
 import * as codecImpl from "./codecimpl";
 import { InitData, Normalize } from "./normalize";
-import { bnsFromOrToTag, bnsNonceTag } from "./tags";
+import { bnsFromOrToTag, bnsNonceTag, bnsSwapQueryTags } from "./tags";
 import { Decoder, Keyed, Result } from "./types";
-import { arraysEqual, bucketKey, hashIdentifier, indexKey, isSwapOffer, isSwapRelease } from "./util";
+import { arraysEqual, hashIdentifier, isSwapOffer, isSwapRelease } from "./util";
 
 // onChange returns a filter than only passes when the
 // value is different than the last one
@@ -62,27 +62,6 @@ function onChange<T>(): (val: T) => boolean {
 // same interface we have with the BCP protocol.
 // We can embed in iov-core process or use this in a BCP-relay
 export class Client implements BcpAtomicSwapConnection {
-  public static swapQueryTags(query: BcpSwapQuery, set = true): Tag {
-    let binKey: Uint8Array;
-    const bucket = "esc";
-    if (isQueryBySwapId(query)) {
-      binKey = Uint8Array.from([...bucketKey(bucket), ...query.swapid]);
-    } else if (isQueryBySwapSender(query)) {
-      binKey = Uint8Array.from([...indexKey(bucket, "sender"), ...query.sender]);
-    } else if (isQueryBySwapRecipient(query)) {
-      binKey = Uint8Array.from([...indexKey(bucket, "recipient"), ...query.recipient]);
-    } else {
-      // if (isQueryBySwapHash(query))
-      binKey = Uint8Array.from([...indexKey(bucket, "arbiter"), ...hashIdentifier(query.hashlock)]);
-    }
-
-    const key = Encoding.toHex(binKey).toUpperCase();
-    // "s" for set, "d" for delete.... we need to watch both changes to be clear
-    // But if we return two tags here, that would AND not OR
-    const value = set ? "s" : "d";
-    return { key, value };
-  }
-
   public static async connect(url: string): Promise<Client> {
     const tm = await TendermintClient.connect(url);
     const initData = await this.initialize(tm);
@@ -228,9 +207,9 @@ export class Client implements BcpAtomicSwapConnection {
   public async getSwap(query: BcpSwapQuery): Promise<BcpQueryEnvelope<BcpAtomicSwap>> {
     // we need to combine them all to see all transactions that affect the query
     const setTxs = await this.searchTx({
-      tags: [Client.swapQueryTags(query, true)],
+      tags: [bnsSwapQueryTags(query, true)],
     });
-    const delTxs = await this.searchTx({ tags: [Client.swapQueryTags(query, false)] });
+    const delTxs = await this.searchTx({ tags: [bnsSwapQueryTags(query, false)] });
 
     // tslint:disable-next-line:readonly-array
     const offers: OpenSwap[] = setTxs.filter(isSwapOffer).map(Normalize.swapOfferFromTx(this.initData));
@@ -256,8 +235,8 @@ export class Client implements BcpAtomicSwapConnection {
   // this includes an open swap beind claimed/expired as well as a new matching swap being offered
   public watchSwap(query: BcpSwapQuery): Stream<BcpAtomicSwap> {
     // we need to combine them all to see all transactions that affect the query
-    const setTxs = this.liveTx({ tags: [Client.swapQueryTags(query, true)] });
-    const delTxs = this.liveTx({ tags: [Client.swapQueryTags(query, false)] });
+    const setTxs = this.liveTx({ tags: [bnsSwapQueryTags(query, true)] });
+    const delTxs = this.liveTx({ tags: [bnsSwapQueryTags(query, false)] });
 
     const offers: Stream<OpenSwap> = setTxs.filter(isSwapOffer).map(Normalize.swapOfferFromTx(this.initData));
