+++ packages/iov-bns/types/client.d.ts
@@ -5,7 +5,6 @@ import { ChainId, PostableBytes, Tag, TxQuery } from "@iov/tendermint-types";
 import { InitData } from "./normalize";
 import { Result } from "./types";
 export declare class Client implements BcpAtomicSwapConnection {
-    static swapQueryTags(query: BcpSwapQuery, set?: boolean): Tag;
     static connect(url: string): Promise<Client>;
     protected static initialize(tmClient: TendermintClient): Promise<InitData>;
     protected readonly tmClient: TendermintClient;
